Revert "Fixes for TestingSyncUpForm"
This reverts commit 5bece14c2a5ae2afda40b57eccc3fab3adb8bdee.
