+++ Sources/ComposableArchitecture/Documentation.docc/Tutorials/BuildingSyncUps/03-SyncUpForm/TestingSyncUpForm-01-code-0004-previous.swift
@@ -6,6 +6,6 @@ import XCTest
 class SyncUpFormTests: XCTestCase {
   @MainActor
   func testRemoveAttendee() async {
-
+    // ...
   }
 }
